requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8,<3

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0 

# AI
google-genai>=1.24.0
